requirements.txt
# Core dependencies
graphiti-core==0.11.6
neo4j>=5.28.0
openai>=1.0.0
python-dotenv>=1.0.0

# Pydantic AI for Q&A agent
pydantic-ai==0.2.9
pydantic-ai-slim==0.2.9
rich==14.0.0

# File parsing
PyPDF2>=3.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
